% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EADMS_Functions.R
\name{Compute_AE}
\alias{Compute_AE}
\title{Compute_AE}
\usage{
Compute_AE(AE)
}
\arguments{
\item{AE}{This is the data received from exporting the AE ad-hoc in Infinite Campus}
}
\description{
This function performs all computations needed for AE, assuming you loaded the data in correctly.
}
\examples{
Compute_AE(AE)
}
\keyword{AE}
\keyword{CUSD}

