% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EADMS_Functions.R
\name{Compute_Elementary}
\alias{Compute_Elementary}
\title{Compute_Elementary}
\usage{
Compute_Elementary(Elem)
}
\arguments{
\item{Elem}{This is the data received from exporting the Elementary ad-hoc in Infinite Campus}
}
\description{
This function performs all computations needed for Elem, assuming you loaded the data in correctly.
}
\examples{
Compute_Elementary(Elem)
}
\keyword{CUSD}
\keyword{Elem}

