% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EADMS_Functions.R
\name{Compute_Secondary}
\alias{Compute_Secondary}
\title{Compute_Secondary}
\usage{
Compute_Secondary(Secondary)
}
\arguments{
\item{Secondary}{This is the data received from exporting the Secondary ad-hoc in Infinite Campus}
}
\description{
This function performs all computations needed for Secondary, assuming you loaded the data in correctly.
}
\examples{
Compute_Secondary(Secondary)
}
\keyword{CUSD}
\keyword{Secondary}

