% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Welcome.R
\name{welcome}
\alias{welcome}
\title{A Welcome Function}
\usage{
welcome(district = TRUE)
}
\arguments{
\item{district}{This is the default parameter. Defaults to TRUE.}
}
\description{
This function welcomes you to CUSD.
}
\examples{
welcome()
}
\keyword{TEST_TEST}

